src/doc-qa/api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-cloud-aiplatform>=1.38.0
google-cloud-bigquery>=3.13.0
google-cloud-secret-manager>=2.18.0
elasticsearch>=8.12.0
